requirements.txt
streamlit
telethon
qrcode
pillow
asyncio